refactor: extract shared and query API crates
