+++ src/api/src/handler/http/request/mod.rs
@@ -13,9 +13,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-use serde::{Deserialize, Serialize};
-use utoipa::ToSchema;
-
 #[cfg(feature = "enterprise")]
 pub mod action_server;
 pub mod actions;
@@ -52,7 +49,7 @@ pub mod pipeline;
 pub mod pipelines;
 #[cfg(feature = "profiling")]
 pub mod profiling;
-pub mod promql;
+pub use openobserve_api_query::{promql, search, traces};
 #[cfg(feature = "enterprise")]
 pub mod providers;
 pub mod ratelimit;
@@ -63,65 +60,14 @@ pub mod rum;
 pub mod score_configs;
 #[cfg(feature = "enterprise")]
 pub mod scorers;
-pub mod search;
 pub mod service_accounts;
 pub mod service_streams;
 pub mod short_url;
 pub mod sourcemaps;
 pub mod status;
 pub mod stream;
-pub mod traces;
 pub mod users;
 
+pub use openobserve_api_shared::request::{BulkDeleteRequest, BulkDeleteResponse};
+
 pub use crate::common::meta::http::{CONTENT_TYPE_JSON, CONTENT_TYPE_PROTO};
-
-// these are the common bulk delete req/res structs
-
-#[derive(Clone, Debug, Serialize, Deserialize, ToSchema)]
-pub struct BulkDeleteRequest {
-    pub ids: Vec<String>,
-}
-
-#[derive(Default, Serialize, ToSchema)]
-pub struct BulkDeleteResponse {
-    pub successful: Vec<String>,
-    pub unsuccessful: Vec<String>,
-    pub err: Option<String>,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_bulk_delete_response_default() {
-        let resp = BulkDeleteResponse::default();
-        assert!(resp.successful.is_empty());
-        assert!(resp.unsuccessful.is_empty());
-        assert!(resp.err.is_none());
-    }
-
-    #[test]
-    fn test_bulk_delete_request_roundtrip() {
-        let req = BulkDeleteRequest {
-            ids: vec!["id1".to_string(), "id2".to_string()],
-        };
-        let json = serde_json::to_string(&req).unwrap();
-        let deserialized: BulkDeleteRequest = serde_json::from_str(&json).unwrap();
-        assert_eq!(deserialized.ids.len(), 2);
-        assert_eq!(deserialized.ids[0], "id1");
-    }
-
-    #[test]
-    fn test_bulk_delete_response_serializes() {
-        let resp = BulkDeleteResponse {
-            successful: vec!["ok1".to_string()],
-            unsuccessful: vec!["fail1".to_string()],
-            err: Some("some error".to_string()),
-        };
-        let json = serde_json::to_string(&resp).unwrap();
-        assert!(json.contains("ok1"));
-        assert!(json.contains("fail1"));
-        assert!(json.contains("some error"));
-    }
-}
