+++ src/api_query/src/lib.rs
@@ -5,7 +5,7 @@
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
 //
-// This program is distributed in the hope that it will be useful
+// This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU Affero General Public License for more details.
@@ -13,8 +13,15 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-pub mod auth;
+//! Search, PromQL, and trace HTTP APIs.
+
+#![recursion_limit = "256"]
+
 pub use openobserve_api_shared::extractors;
-pub mod models;
-pub mod request;
-pub mod router;
+#[cfg(feature = "enterprise")]
+pub use openobserve_core::cipher;
+pub use openobserve_core::{common, service};
+
+pub mod promql;
+pub mod search;
+pub mod traces;
