+++ src/api_query/src/search/query_manager.rs
