+++ src/api_query/src/traces/schema_compat.rs
