+++ src/api_query/src/traces/session.rs
@@ -33,9 +33,8 @@ use crate::{
             http::{get_or_create_trace_id, get_use_cache_from_request},
         },
     },
-    handler::http::{
-        extractors::Headers, request::search::error_utils::map_error_to_http_response,
-    },
+    extractors::Headers,
+    search::error_utils::map_error_to_http_response,
     service::{search as SearchService, traces},
 };
 
@@ -133,7 +132,7 @@ pub async fn get_latest_sessions(
             let user: config::meta::user::User = get_user(Some(&org_id), user_id).await.unwrap();
             let stream_type_str = StreamType::Traces.as_str();
 
-            if !crate::handler::http::auth::validator::check_permissions(
+            if !crate::service::authz::check_permissions(
                 user_id,
                 AuthExtractor {
                     auth: "".to_string(),
@@ -699,7 +698,7 @@ pub async fn get_session_details(
             let user: config::meta::user::User = get_user(Some(&org_id), user_id).await.unwrap();
             let stream_type_str = StreamType::Traces.as_str();
 
-            if !crate::handler::http::auth::validator::check_permissions(
+            if !crate::service::authz::check_permissions(
                 user_id,
                 AuthExtractor {
                     auth: "".to_string(),
